+++ OneBlink.SDK/models/FormsAppEnvironmentsSearchResult.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+namespace OneBlink.SDK.Model
+{
+    public class FormsAppEnvironmentsSearchResult : SearchResult
+    {
+        public List<FormsAppEnvironment> formsAppEnvironments
+        {
+            get; set;
+        }
+    }
+}