+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 -   `emailTemplate` to `FormSubmissionEventConfiguration`
+-   Added `FormsAppEnvironmentsClient`
 
 ## 2.5.3 (2021-10-05)